home page UI, card flip, componentization
