+++ web/src/app/home/leaderboard/page.tsx
@@ -3,7 +3,7 @@ import LeaderboardView from "@/components/leaderboard/leaderboard-view";
 
 export default async function LeaderboardPage() {
   return (
-    <div className="w-full h-full overflow-y-auto flex flex-col gap-5 bg-gradient-to-b from-[#1E1E1E] via-[#050505] to-black text-white pt-[22px]">
+    <div className="w-full min-h-screen overflow-y-auto flex flex-col flex-1 gap-5 bg-gradient-to-b from-[#1E1E1E] via-[#050505] to-black text-white pt-5">
       <Navbar />
       <LeaderboardView />
     </div>
