+++ web/src/app/layout.tsx
@@ -1,29 +1,12 @@
 "use client";
 
+import { Suspense } from "react";
 import "./globals.css";
-import {
-  SolanaProvider,
-  WalletButton,
-} from "@/components/solana/solana-provider";
+import { SolanaProvider } from "@/components/solana/solana-provider";
 import { ReactQueryProvider } from "./react-query-provider";
 import { TRPCReactProvider } from "@/trpc/react";
-import { SessionProvider, useSession } from "next-auth/react";
+import { SessionProvider } from "next-auth/react";
 import { Toaster } from "react-hot-toast";
-import { Suspense, useEffect, useState } from "react";
-import { usePathname } from "next/navigation";
-import { useRouter } from "next/navigation";
-import { capsule } from "@/lib/capsule";
-import { Button } from "@/components/ui/button";
-import { PublicKey } from "@solana/web3.js";
-import { IconCopy } from "@tabler/icons-react";
-import Link from "next/link";
-import { useWallet } from "@solana/wallet-adapter-react";
-import { shortenAddress } from "@/lib/utils";
-
-const links: { label: string; path: string }[] = [
-  { label: "Account", path: "/account" },
-  { label: "Web Program", path: "/web" },
-];
 
 export default function RootLayout({
   children,
@@ -32,27 +15,13 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en" className="bg-background">
-      <body className="h-screen overflow-hidden bg-[url('/background.png')] bg-cover bg-center bg-fixed">
+      <body className="h-screen bg-[url('/background.png')] bg-cover bg-center bg-fixed">
         {/* Dark overlay */}
         <SessionProvider>
           <ReactQueryProvider>
             <SolanaProvider>
               <TRPCReactProvider>
                 <Toaster />
-                <div className="navbar  text-neutral-content">
-                  <div className="flex-1">
-                    <Link
-                      className="btn btn-ghost normal-case text-xl"
-                      href="/"
-                    >
-                      {/* <img className="h-4 md:h-6" alt="Logo" src="/logo.png" /> */}
-                      TRADETALK
-                    </Link>
-                  </div>
-                  <div className="flex-none space-x-2">
-                    <AccountButtons />
-                  </div>
-                </div>
                 <Suspense
                   fallback={
                     <div className="text-center my-32">
@@ -60,7 +29,7 @@ export default function RootLayout({
                     </div>
                   }
                 >
-                  <main className="relative  max-w-screen min-h-screen bg-background">
+                  <main className="relative max-w-screen min-h-screen">
                     {children}
                   </main>
                 </Suspense>
@@ -70,69 +39,5 @@ export default function RootLayout({
         </SessionProvider>
       </body>
     </html>
-  );
-}
-
-function AccountButtons() {
-  const [isActive, setIsActive] = useState(false);
-  const router = useRouter();
-  const pathname = usePathname();
-  const { publicKey } = useWallet();
-  const { data: session } = useSession();
-
-  useEffect(() => {
-    capsule.isSessionActive().then(setIsActive);
-  }, [setIsActive, session, publicKey]);
-
-  if (publicKey) {
-    return <WalletButton />;
-  }
-
-  if (!session) {
-    return (
-      <Button
-        onClick={() => {
-          console.log("redirecting to signin");
-          router.push(
-            `/auth/signin?callbackUrl=${encodeURIComponent(pathname)}`
-          );
-        }}
-      >
-        Sign in
-      </Button>
-    );
-  }
-
-  if (!isActive) {
-    return (
-      <Button
-        onClick={() => {
-          console.log("redirecting to signin");
-          router.push(
-            `/auth/signin?callbackUrl=${encodeURIComponent(pathname)}`
-          );
-        }}
-      >
-        Sign in
-      </Button>
-    );
-  }
-  try {
-    const pk = new PublicKey(capsule.getAddress()!);
-  } catch (e) {
-    return null;
-  }
-
-  const pk = new PublicKey(capsule.getAddress()!);
-  return (
-    <div
-      className="btn btn-primary flex flex-row gap-2 h-full"
-      onClick={() => {
-        navigator.clipboard.writeText(pk.toBase58());
-      }}
-    >
-      <IconCopy />
-      {shortenAddress(pk)}
-    </div>
   );
 }
