+++ web/src/components/home/player-card-back.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { usePlayerMarketCardStore } from "@/lib/zustand";
+import { Trade2 } from "../web/web-ui";
+
+export default function PlayerCardBack() {
+  const { selectedOrderType, activePlayerMarket } = usePlayerMarketCardStore();
+
+  return <Trade2 />;
+}
