+++ web/src/lib/zustand.tsx
@@ -1,5 +1,10 @@
 import { create } from "zustand";
 
+export enum CardView {
+  FRONT = "front",
+  BACK = "back",
+}
+
 interface FiltersState {
   activeFilters: string[];
   setActiveFilters: (filters: string[]) => void;
@@ -9,9 +14,13 @@ interface FiltersState {
   setShowBookmarked: (show: boolean) => void;
 }
 
-interface ActivePlayerMarketState {
+interface PlayerMarketCardState {
   activePlayerMarket: string;
   setActivePlayerMarket: (market: string) => void;
+  selectedOrderType: "buy" | "sell";
+  setSelectedOrderType: (orderType: "buy" | "sell") => void;
+  cardView: CardView;
+  setCardView: (cardView: CardView) => void;
 }
 
 export const useFiltersStore = create<FiltersState>()((set) => ({
@@ -23,9 +32,13 @@ export const useFiltersStore = create<FiltersState>()((set) => ({
   setShowBookmarked: (show) => set({ showBookmarked: show }),
 }));
 
-export const useActivePlayerMarketStore = create<ActivePlayerMarketState>()(
+export const usePlayerMarketCardStore = create<PlayerMarketCardState>()(
   (set) => ({
     activePlayerMarket: "",
     setActivePlayerMarket: (market) => set({ activePlayerMarket: market }),
+    selectedOrderType: "buy",
+    setSelectedOrderType: (orderType) => set({ selectedOrderType: orderType }),
+    cardView: CardView.FRONT,
+    setCardView: (cardView) => set({ cardView }),
   })
 );
